[canvaskit] Add ImageInfo to Texture APIs
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/470736

+++ modules/canvaskit/tests/core.spec.js
@@ -1290,7 +1290,7 @@ describe('Core canvas behavior', () => {
         if (!CanvasKit.gpu) {
             return;
         }
-        // This creates and draws an Image that is 1 pixel wide, 4 pixels tall with
+        // This creates and draws an Unpremul Image that is 1 pixel wide, 4 pixels tall with
         // the colors listed below.
         const pixels = Uint8Array.from([
             255,   0,   0, 255, // opaque red
@@ -1298,7 +1298,12 @@ describe('Core canvas behavior', () => {
               0,   0, 255, 255, // opaque blue
             255,   0, 255, 100, // transparent purple
         ]);
-        const img = surface.makeImageFromTextureSource(pixels, 1, 4);
+        const img = surface.makeImageFromTextureSource(pixels, {
+              'width': 1,
+              'height': 4,
+              'alphaType': CanvasKit.AlphaType.Unpremul,
+              'colorType': CanvasKit.ColorType.RGBA_8888,
+            });
         canvas.drawImage(img, 1, 1, null);
 
         const info = img.getImageInfo();
@@ -1312,6 +1317,36 @@ describe('Core canvas behavior', () => {
         expect(CanvasKit.ColorSpace.Equals(cs, CanvasKit.ColorSpace.SRGB)).toBeTruthy();
 
         cs.delete();
+        img.delete();
+    });
+
+    gm('makeImageFromTextureSource_PremulTypedArray', (canvas, _, surface) => {
+        if (!CanvasKit.gpu) {
+            return;
+        }
+        // This creates and draws an Unpremul Image that is 1 pixel wide, 4 pixels tall with
+        // the colors listed below.
+        const pixels = Uint8Array.from([
+            255,   0,   0, 255, // opaque red
+              0, 255,   0, 255, // opaque green
+              0,   0, 255, 255, // opaque blue
+            100,   0, 100, 100, // transparent purple
+        ]);
+        const img = surface.makeImageFromTextureSource(pixels, {
+              'width': 1,
+              'height': 4,
+              'alphaType': CanvasKit.AlphaType.Premul,
+              'colorType': CanvasKit.ColorType.RGBA_8888,
+            });
+        canvas.drawImage(img, 1, 1, null);
+
+        const info = img.getImageInfo();
+        expect(info).toEqual({
+          'width': 1,
+          'height': 4,
+          'alphaType': CanvasKit.AlphaType.Premul,
+          'colorType': CanvasKit.ColorType.RGBA_8888,
+        });
         img.delete();
     });
 
@@ -1361,6 +1396,29 @@ describe('Core canvas behavior', () => {
               'alphaType': CanvasKit.AlphaType.Unpremul,
               'colorType': CanvasKit.ColorType.RGBA_8888,
             });
+            img.delete();
+        });
+    });
+
+    gm('MakeLazyImageFromTextureSource_imageInfo', (canvas) => {
+        if (!CanvasKit.gpu) {
+            return;
+        }
+        // This makes an offscreen <img> with the provided source.
+        const imageEle = new Image();
+        imageEle.src = '/assets/mandrill_512.png';
+
+        // We need to wait until the image is loaded before the texture can use it. For good
+        // measure, we also wait for it to be decoded.
+        return imageEle.decode().then(() => {
+            const img = CanvasKit.MakeLazyImageFromTextureSource(imageEle, {
+              'width': 400,
+              'height': 400,
+              'alphaType': CanvasKit.AlphaType.Premul,
+              'colorType': CanvasKit.ColorType.RGBA_8888,
+            });
+            canvas.drawImage(img, 20, 20, null);
+
             img.delete();
         });
     });
